% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colPalette.R
\name{colPalette}
\alias{colPalette}
\title{colPalette}
\usage{
colPalette(palette)
}
\arguments{
\item{palette}{A character string specifying the name of the colour palette.
The palette must be present in either 'taylor_palettes' or 'radiohead_palettes'.}
}
\value{
A colorRampPalette function with the specified colour palette.
}
\description{
Create a Color Palette. This function generates a custom
color palette based on the specified 'palette' name.
The color palettes are sourced from two
predefined lists: 'taylor_palettes' and 'radiohead_palettes'.
}
\examples{
my_palette <- colPalette(palette = "evermore")
colours <- my_palette(5)
print(colors)
}
