% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtension.R
\name{getExtension}
\alias{getExtension}
\title{getExtension}
\usage{
getExtension(file)
}
\arguments{
\item{file}{A character string containing the file path.}
}
\value{
A character vector containing the file extension without the dot.
}
\description{
Get the file extension of a given file.
This function takes a file path as input and returns the file extension without the dot.
}
\examples{
getExtension("example.txt")

}
