% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupCols.R
\name{groupCols}
\alias{groupCols}
\title{groupCols}
\usage{
groupCols(hex_colors, n_clusters = 5)
}
\arguments{
\item{hex_colors}{A character vector of hex color values.}

\item{n_clusters}{The number of clusters (groups) to use in the k-means clustering. Default is 5.}
}
\value{
A data frame with the original hex colors and their assigned group labels.
}
\description{
Group Hex Colors. This function takes a vector of hex color values
and groups them using k-means clustering in the RGB color space.
It returns a data frame with the original hex colors and their assigned group labels.
}
\examples{
hex_colors <- c("#FF0000", "#00FF00", "#0000FF", "#FFFF00", "#FF00FF")
groupCols(hex_colors, n_clusters = 2)
}
