% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioheadPals.R
\name{radiohead_pal}
\alias{radiohead_pal}
\title{Function to interpolate a color palette}
\usage{
radiohead_pal(palette = "pabloHoney", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tswift_palettes}

\item{reverse}{Boolean true if palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
A vector of colors
}
\description{
Function to interpolate a color palette
}
