% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioheadPals.R
\docType{data}
\name{radiohead_palettes}
\alias{radiohead_palettes}
\title{list of radiohead colour palettes}
\format{
An object of class \code{list} of length 9.
}
\usage{
radiohead_palettes
}
\description{
list of radiohead colour palettes
}
\keyword{datasets}
