% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioheadPals.R
\name{scale_color_radiohead}
\alias{scale_color_radiohead}
\title{Color scale for Radiohead colors}
\usage{
scale_color_radiohead(
  palette = "pabloHoney",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in Radiohead_palettes}

\item{discrete}{Boolean if color aesthetic is discrete}

\item{reverse}{Boolean indicating whether palette should be reversed}

\item{...}{Additional arguments used to discrete_scale() or scale_fill_gradientn()
to automatically interpolate between colours.}
}
\value{
No return value. Called for side effects
}
\description{
Color scale for Radiohead colors
}
\examples{
library(ggplot2)
data <- data.frame(c = LETTERS[1:3],x = c(1,5,7),y = c(5,9,13))
ggplot(data, aes(x,y,color = c)) + geom_point() + scale_color_radiohead()
}
