% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylorPals.R
\docType{data}
\name{taylor_palettes}
\alias{taylor_palettes}
\title{list of Taylor swift colour palettes}
\format{
An object of class \code{list} of length 10.
}
\usage{
taylor_palettes
}
\description{
list of Taylor swift colour palettes
}
\keyword{datasets}
