% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopCol.R
\name{excludeCols}
\alias{excludeCols}
\title{excludeCols}
\usage{
excludeCols(data, customExclude = NULL)
}
\arguments{
\item{data}{Data frame of colours}

\item{customExclude}{Character vector. Optional vector of custom color codes in HEX format to be excluded.}
}
\value{
No return value, called for side effects.
}
\description{
A non-exhaustive list of white and black colour shades. For use when setting \code{exclude = TRUE}
in the \code{getTopCol} function. Setting  \code{exclude = TRUE} when calling the \code{getTopCol}
function will exclude the colours form the results.
}
